Don't use GGUF but convert T5 directly from Hugging Face (#967)
We don't want the stack to depend on the conversion tool from the
lamma.cpp repo. Also the conversion to GGUF would not convert all
tensors to bf16, but leave some in f32. We would like to control that
ourselves if needed.

This change makes any previously generated IRPA files obsolete.