Llm use task inputs (#2195)
## Description

- Queue up and invoke on `LlmTaskInputs`
- Plumb through a `--use_chunked_prefill` arg as a feature flag to use
while developing

This is intended as a precursor to `chunked prefill`. The idea is:

"high level idea - once the batcher receives its requests it processes
them to generate a bunch of task inputs that need to be executed then
places it in a following queue. This is the queue that gets strobed /
processed on.

For the current world we just generate full prefill requests and jam it
in (piece of cake). For chunking we generate a set of input requests
that are chunked instead. Then we are only thinking about how to chunk
work on reception instead of trying to process and roundtrip requests."

The `BatcherProcess` has an [infeed
queue](https://github.com/nod-ai/shark-ai/blob/main/shortfin/python/shortfin_apps/utils.py#L597)
that receives an `LlmInferenceExecRequest`.

From there, instead of submitting an `LlmInferenceExecRequest` to the
`scheduler` and `invocation` layer, we generate the `LlmTaskInputs` from
that request and work with those instead.

The follow up idea would be that we would be able to store the list of
chunks somewhere, either a `ChunkedScheduler` or the batcher`, and just
generically queue up the next chunk (`LlmTaskInput`) from that request
when it's done. That would require work on the `ChunkedScheduler`. This
setups the current layer, so that it should fit in with it pretty well.

The end goal is this design:
<img width="846" height="821" alt="chunked_prefill_scheduler drawio"
src="https://github.com/user-attachments/assets/5026f3bc-6d64-4752-a5ec-08bfec433d34"
/>